refactor(usage): split tracker.ts into types/windows/summary modules
